btc converter, publisher, and indexer unit tests

+++ db/migrations/00044_create_btc_tx_outputs_table.sql
@@ -3,7 +3,7 @@ CREATE TABLE btc.tx_outputs (
   id            SERIAL PRIMARY KEY,
 	tx_id         INTEGER NOT NULL REFERENCES btc.transaction_cids (id) ON DELETE CASCADE DEFERRABLE INITIALLY DEFERRED,
 	index         INTEGER NOT NULL,
-	value         INTEGER NOT NULL,
+	value         BIGINT NOT NULL,
 	pk_script     BYTEA NOT NULL,
 	script_class  INTEGER NOT NULL,
 	addresses     VARCHAR(66)[],
